86bzxxmxk Move `ContactMapper` into `mappers` folder
