+++ src/modules/contacts/contacts.module.ts
@@ -4,7 +4,7 @@ import { ContactsController } from './contacts.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Contact } from './entities/contact.entity';
 import { Board } from '../boards/entities/board.entity';
-import { ContactMapper } from './contacts.mapper';
+import { ContactMapper } from './mappers/contacts.mapper';
 import { JobApplication } from '../job-applications/entities/job-application.entity';
 import { ContactEmail } from './entities/contact-emails.entity';
 import { ContactPhone } from './entities/contact-phones.entity';
