+++ src/modules/contacts/mappers/contacts.mapper.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
-import { Contact } from './entities/contact.entity';
-import { ContactDto } from './dtos/contact.dto';
-import { CreateContactDto } from './dtos/create-contact.dto';
-import { Board } from '../boards/entities/board.entity';
-import { BoardMapper } from '../boards/boards.mapper';
-import { JobApplicationMapper } from '../job-applications/job-applications.mapper';
+import { Contact } from '../entities/contact.entity';
+import { ContactDto } from '../dtos/contact.dto';
+import { CreateContactDto } from '../dtos/create-contact.dto';
+import { Board } from '../../boards/entities/board.entity';
+import { BoardMapper } from '../../boards/boards.mapper';
+import { JobApplicationMapper } from '../../job-applications/job-applications.mapper';
 
 @Injectable()
 export class ContactMapper {
